+++ src/base/routes/main.js
@@ -527,11 +527,6 @@ export default [
         ],
       },
       {
-        // TODO: legacy redirect, can be removed in some months
-        path: 'place/:placeId/activities/:activityId/detail',
-        redirect: 'place/:placeId/activities/:activityId/detail',
-      },
-      {
         name: 'activityDetail',
         path: 'place/:placeId/activities/:activityId/detail',
         meta: {