+++ src/messages/components/ConversationMessage.story.js
@@ -1,7 +1,7 @@
 import { storiesOf } from '@storybook/vue'
 
 import ConversationMessage from './ConversationMessage'
-import { storybookDefaults as defaults } from '>/helpers'
+import { createDatastore, storybookDefaults as defaults } from '>/helpers'
 import * as factories from '>/enrichedFactories'
 
 const range = n => [...Array(n).keys()]
@@ -12,8 +12,17 @@ const message = factories.makeMessage({
 })
 const thread = factories.makeThread()
 
+const store = createDatastore({
+  users: {
+    getters: {
+      byCurrentGroup: () => [],
+    },
+  },
+})
+
 storiesOf('ConversationMessage', module)
   .add('default', () => defaults({
+    store,
     render: h => h(ConversationMessage, {
       props: {
         message,
@@ -21,6 +30,7 @@ storiesOf('ConversationMessage', module)
     }),
   }))
   .add('edited', () => defaults({
+    store,
     render: h => h(ConversationMessage, {
       props: {
         message: {
@@ -31,6 +41,7 @@ storiesOf('ConversationMessage', module)
     }),
   }))
   .add('via email', () => defaults({
+    store,
     render: h => h(ConversationMessage, {
       props: {
         message: {
@@ -41,6 +52,7 @@ storiesOf('ConversationMessage', module)
     }),
   }))
   .add('with unread replies', () => defaults({
+    store,
     render: h => h(ConversationMessage, {
       props: {
         message: thread,
@@ -48,6 +60,7 @@ storiesOf('ConversationMessage', module)
     }),
   }))
   .add('with replies', () => defaults({
+    store,
     render: h => h(ConversationMessage, {
       props: {
         message: {
@@ -61,6 +74,7 @@ storiesOf('ConversationMessage', module)
     }),
   }))
   .add('with one image', () => defaults({
+    store,
     render: h => h(ConversationMessage, {
       props: {
         message: {
@@ -73,6 +87,7 @@ storiesOf('ConversationMessage', module)
     }),
   }))
   .add('with multiple images', () => defaults({
+    store,
     render: h => h(ConversationMessage, {
       props: {
         message: {
@@ -83,6 +98,7 @@ storiesOf('ConversationMessage', module)
     }),
   }))
   .add('slim', () => defaults({
+    store,
     render: h => h(ConversationMessage, {
       props: {
         message,